requirements.txt
accelerate==0.23.0
albumentations==1.3.1
annotated-types==0.6.0
anyio==3.7.1
arrow==1.3.0
arxiv==1.4.8
backoff==2.2.1
beautifulsoup4==4.12.2
bibtexparser @ git+https://github.com/sciunto-org/python-bibtexparser@770eb001349dcad48149bc101f07cfc88282d8b3
bitsandbytes==0.41.1
blessed==1.20.0
bs4==0.0.1
certifi==2023.7.22
chardet==5.2.0
charset-normalizer==3.2.0
contourpy==1.1.1
croniter==1.4.1
cycler==0.12.0
datasets==2.14.5
dateutils==0.6.12
deepdiff==6.6.0
dill==0.3.7
EasyProcess==1.1
efficiency @ git+https://github.com/zhijing-jin/efficiency.git@dcee338bd0f0cb0d51f5be268b0c58c239c2c309
entrypoint2==1.1
evaluate==0.4.3
fastapi==0.103.2
feedparser==6.0.10
filelock==3.12.4
fonttools==4.43.0
fsspec==2023.6.0
h11==0.14.0
html2text==2020.1.16
huggingface-hub==0.17.3
imageio==2.31.5
importlib-resources==6.1.0
inquirer==3.1.3
itsdangerous==2.1.2
jeepney==0.8.0
Jinja2==3.1.2
joblib==1.3.2
kiwisolver==1.4.5
langchain==0.3.4
langchain-anthropic=0.2.3
lazy_loader==0.3
Levenshtein==0.22.0
lightning==2.0.9.post0
lightning-cloud==0.5.39
lightning-utilities==0.9.0
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.0
mdurl==0.1.2
MouseInfo==0.1.3
mpmath==1.3.0
mss==9.0.1
multiprocess==0.70.15
munch==4.0.0
networkx==3.1
nltk==3.8.1
numpy==1.26.0
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.2.140
nvidia-nvtx-cu12==12.1.105
openai==0.28.1
opencv-python-headless==4.8.1.78
ordered-set==4.1.0
orjson==3.9.7
pandas==2.1.1
Pillow==10.0.1
protobuf==3.20.3
pyarrow==13.0.0
PyAutoGUI==0.9.54
pydantic==2.1.1
pydantic_core==2.4.0
PyGetWindow==0.0.9
pylatexenc==2.10
PyMsgBox==1.0.9
pyparsing==3.1.1
pypdf==3.16.3
pypdfium2==4.20.0
pyperclip==1.8.2
PyRect==0.2.0
pyscreenshot==3.1
PyScreeze==0.1.29
python-editor==1.0.4
python-Levenshtein==0.22.0
python-multipart==0.0.6
python3-xlib==0.15
pytorch-lightning==2.0.9.post0
pytweening==1.0.7
pytz==2023.3.post1
PyYAML==6.0.1
qudida==0.0.4
rapidfuzz==3.4.0
readchar==4.0.5
regex==2023.10.3
rich==13.6.0
ruamel.yaml==0.17.35
ruamel.yaml.clib==0.2.8
safetensors==0.4.0
scikit-image==0.22.0
scikit-learn==1.3.1
scipy==1.11.3
sconf==0.2.5
sentencepiece==0.1.99
sentence-transformers==3.2.1
sgmllib3k==1.0.0
sniffio==1.3.0
soupsieve==2.5
starlette==0.27.0
starsessions==1.3.0
sympy==1.12
threadpoolctl==3.2.0
tifffile==2023.9.26
timm==0.5.4
tokenizers==0.14.1
torch==2.1.0
torch-tb-profiler==0.4.3
torchmetrics==1.2.0
torchvision==0.16.0
tqdm==4.66.1
transformers==4.34.0
triton==2.1.0
types-python-dateutil==2.8.19.14
tzdata==2023.3
urllib3==2.0.5
uvicorn==0.23.2
websocket-client==1.6.4
websockets==11.0.3
Werkzeug==3.0.0
xxhash==3.4.1